Add files via upload
There are totally 7 files in this repository and the main purpose of this graduation project is to recognize characters and digits in the given handwritten text and to digitalize them. Actually, there are two forms of the project above. The first consisted of HandwrittenDigitRecognition, SVCmodel, KNNmodel, and dataset. The second one consisted of HandwrittenDigitRecognitionSecond, supportVectorClassifierModel, and datasetEdit files. Additionally, these files are not a whole project and also are preliminary work for the graduation project. In this file, first of all, I create images from Paint and load them into the system to be read. Therefore, I created a dataset and model according to these images. As for controlling the accuracy model, I tried to improve it. At the end of the program, I created a simple interface in Windows to check whether the given number from the user in Paint is true or not. Totally, there is an accuracy worth 76%. I will be publishing a continuation of the program.